profile corrections, trim, navbar smaller, more videos on main

+++ src/components/registration/RegisterPage.js
@@ -14,6 +14,7 @@ const RegisterPage = () => {
   const [showPasswordError, setShowPasswordError] = useState(false);
   const [showPasswordConfirmError, setShowPasswordConfirmError] = useState(false);
   const [showRegisterSuccessful, setShowRegisterSuccessful] = useState(false);
+  const [showUserAlreadyExistError, setShowUserAlreadyExistError] = useState(false);
   const navigate = useNavigate();
 
   const handleUsernameChange = (event) => {
@@ -34,13 +35,13 @@ const RegisterPage = () => {
   const handleSubmit = (event) => {
     event.preventDefault();
     let valid = true;
-
+  
     // Validate username
     if (!username.trim()) {
       setShowUsernameError(true);
       valid = false;
     }
-
+  
     // Validate password
     const passwordRegex =
       /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[^a-zA-Z0-9]).{6,}$/;
@@ -48,54 +49,63 @@ const RegisterPage = () => {
       setShowPasswordError(true);
       valid = false;
     }
-
+  
     // Validate password confirmation
     if (password !== passwordConfirm) {
       setShowPasswordConfirmError(true);
       valid = false;
     }
-
+  
     if (valid) {
+      // Retrieve existing users from local storage, or start with an empty array
+      const existingUsers = JSON.parse(
+        localStorage.getItem("AllUsers") || "[]"
+      );
+  
+      // Check if the user already exists
+      const userExists = existingUsers.some((user) => user.username === username);
+  
+      if (userExists) {
+        setShowUserAlreadyExistError(true)
+        return;
+      }
+  
       const newUser = {
         username,
         nickname: username,
         password,
       };
-
-      // Retrieve existing users from local storage, or start with an empty array
-      const existingUsers = JSON.parse(
-        localStorage.getItem("AllUsers") || "[]"
-      );
-
+  
       // Add the new user to the array of existing users
       const allUsers = [...existingUsers, newUser];
-
+  
       // Save the updated array of users to local storage
       localStorage.setItem("AllUsers", JSON.stringify(allUsers));
-
+  
       // Clear the form
       setUsername("");
       setPassword("");
       setPasswordConfirm("");
-      setShowRegisterSuccessful (true)
+      setShowRegisterSuccessful(true);
       setTimeout(() => {
-      navigate(`/login`);
-    }, 2000);
+        navigate(`/login`);
+      }, 2000);
     }
   };
+  
 
   return (
     <Container className={styles.profileContainer}
     style={{ marginTop: "50px"}}>
       <Row>
-        <Col   style={{ marginTop: "60px", marginLeft: "500px" }}
+        <Col   style={{ marginTop: "80px", marginLeft: "500px" }}
           md={3}
           className="d-flex flex-column align-items-center">
           <form onSubmit={handleSubmit}>
             <div>
-              <label htmlFor="username">Username:</label>
+              <label style={{color:"rgb(3, 140, 252)"}} htmlFor="username">Username:</label>
               <input
-               style={{ marginLeft: "5px" }}
+               style={{ marginLeft: "5px"}}
                 type="text"
                 id="username"
                 value={username}
@@ -105,7 +115,7 @@ const RegisterPage = () => {
             
             </div>
             <div>
-              <label htmlFor="password">Password:</label>
+              <label style={{color:"rgb(3, 140, 252)"}} htmlFor="password">Password:</label>
               <input
                 type="password"
                 id="password"
@@ -116,7 +126,7 @@ const RegisterPage = () => {
              
             </div>
             <div>
-              <label htmlFor="password-confirm">Confirm Password:</label>
+              <label style={{color:"rgb(3, 140, 252)"}} htmlFor="password-confirm">Confirm Password:</label>
               <input
                 type="password"
                 id="password-confirm"
@@ -129,9 +139,9 @@ const RegisterPage = () => {
 
             <Link style={{color:"rgb(93, 147, 177)" }} to="/login">Already have an account?</Link>
 
-            <Button style={{ margin: "5px" }} type="submit">Register</Button>
+            <Button style={{ margin: "5px",   background:"rgb(63, 63, 68)" }} type="submit">Register</Button>
 
-            <div style={{color:"rgb(93, 147, 177)", position: "absolute"}}>
+            <div style={{color:"rgb(240, 7, 34)", position: "absolute"}}>
               <div>
               {showUsernameError && (
                   <div className="error"> • Please enter a username</div>
@@ -147,9 +157,13 @@ const RegisterPage = () => {
                 )}
               </div>
               {showRegisterSuccessful && (
-                  <div className="error">Registration Successful!
+                  <div style={{ color: 'rgb(79, 194, 102)' }} className="error">Registration Successful!
+                  </div> )}
+                  {showUserAlreadyExistError && (
+                  <div className="error">• This username is already taken!
                   </div> )}
             </div>
+            
 
 
           </form>
